Add files via upload
Application.cpp -> Glut timer is used for animation.
display.cpp -> stepBall() is added. Function updates the location of the ball every time when function is called. Position of the ball has to be updated with "GLFloat ballPosition[3]". For now, ball will bounce around the room continuously. Animation can be changed by changing the initial values of"GLfloat ballVelocity[3] = { 0.0, 0.0, 0.0 };"